move gob to stdlib; test passes w merge rec ver
