+++ barretenberg/cpp/src/barretenberg/stdlib/recursion/goblin/mock_circuits.hpp
@@ -1,11 +1,10 @@
 #include "barretenberg/commitment_schemes/commitment_key.hpp"
 #include "barretenberg/eccvm/eccvm_composer.hpp"
-#include "barretenberg/goblin/goblin.hpp"
-#include "barretenberg/goblin/translation_evaluations.hpp"
 #include "barretenberg/proof_system/circuit_builder/eccvm/eccvm_circuit_builder.hpp"
 #include "barretenberg/proof_system/circuit_builder/goblin_ultra_circuit_builder.hpp"
 #include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
 #include "barretenberg/translator_vm/goblin_translator_composer.hpp"
+#include "barretenberg/translator_vm/translation_evaluations.hpp"
 #include "barretenberg/ultra_honk/ultra_composer.hpp"
 
 namespace barretenberg {
